web_server/requirements.txt
flask
flask-mysql
Werkzeug
Jinja2
MarkupSafe
itsdangerous
click

